fix double insert if dob is null
Handling the date-of-birth as a string now simplifies the whole process and fixes finally the issue with the timezone convertion.
Since playing with java timeobjects always comes with hours, minutes, seconds and timezones but we have only a date this is the best solution for me here.
For more explanations see README.md